requirements.txt
affine==2.4.0
aiobotocore==2.22.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aioitertools==0.12.0
aiosignal==1.3.1
amqp==5.3.1
anyio==4.5.2
apturl==0.5.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
asttokens==3.0.0
async-lru==2.0.4
async-timeout==5.0.1
attrs==25.1.0
babel==2.17.0
backcall==0.2.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
beautifulsoup4==4.13.3
billiard==4.2.1
bleach==6.1.0
blinker==1.8.2
botocore==1.37.3
bounded-pool-executor==0.0.3
Brlapi==0.7.0
cached-property==1.5.1
cachetools==5.5.2
celery==5.5.3
certifi==2025.7.14
cffi==1.17.1
cftime==1.6.4.post1
chardet==3.0.4
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cligj==0.7.2
colorama==0.4.3
comm==0.2.2
command-not-found==0.3
contourpy==1.1.1
cron-descriptor==1.4.5
cryptography==2.8
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.2.16
debugpy==1.8.13
decorator==5.2.1
defer==1.0.6
defusedxml==0.7.1
distro==1.4.0
distro-info==0.23+ubuntu1.1
dj-database-url==3.0.0
django==4.2.23
django-celery-beat==2.8.1
django-cors-headers==4.4.0
django-timezone-field==7.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dlib==20.0.0
docker==4.1.0
docker-compose==1.25.0
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.12.0
earthaccess==0.9.0
earthengine-api==1.1.5
entrypoints==0.3
exceptiongroup==1.2.2
executing==2.2.0
face-recognition==1.3.0
face-recognition-models==0.3.0
fasteners==0.14.1
fastjsonschema==2.21.1
flask==3.0.3
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2025.3.0
future==0.18.2
geographiclib==2.0
geopy==2.4.1
google-ai-generativelanguage==0.1.0
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-storage==3.2.0
google-crc32c==1.5.0
google-generativeai==0.1.0rc1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
h5netcdf==1.1.0
h5py==3.11.0
httpcore==1.0.7
httplib2==0.14.0
httpx==0.28.1
idna==2.8
importlib-metadata==8.5.0
importlib-resources==6.4.5
ipykernel==6.29.5
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.2
jinja2==3.1.6
jmespath==1.0.1
joblib==1.4.2
json5==0.10.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-client==8.6.3
jupyter-core==5.7.2
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter-server==2.14.2
jupyter-server-terminals==0.5.3
jupyterlab==4.3.5
jupyterlab-pygments==0.3.0
jupyterlab-server==2.27.3
keyring==18.0.1
kiwisolver==1.4.7
kombu==5.5.4
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mistune==3.1.2
monotonic==1.5
more-itertools==4.2.0
multidict==6.1.0
multimethod==1.10
mysql-connector-python==9.0.0
nbclient==0.10.1
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
netCDF4==1.7.2
netifaces==0.10.4
notebook==7.3.2
notebook-shim==0.2.4
numpy==1.24.4
oauthlib==3.1.0
olefile==0.46
opencv-python==4.11.0.86
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.2
pandas==2.0.3
pandocfilters==1.5.1
paramiko==2.6.0
parso==0.8.4
pexpect==4.6.0
pickleshare==0.7.5
pillow==10.4.0
pkgutil-resolve-name==1.3.10
platformdirs==4.3.6
pqdm==0.2.0
prometheus-client==0.21.1
prompt-toolkit==3.0.50
propcache==0.2.0
proto-plus==1.26.1
protobuf==4.25.8
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure-eval==0.2.3
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycairo==1.16.2
pycparser==2.22
pycups==1.9.73
pygments==2.19.1
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==3.1.4
pyRFC3339==1.1
pyrsistent==0.15.5
PySocks==1.7.1
python-apt==2.0.1+ubuntu0.20.4.1
python-cmr==0.13.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-debian==0.1.36+ubuntu1.1
python-decouple==3.8
python-json-logger==3.3.0
pytz==2025.2
pyxdg==0.26
PyYAML==5.3.1
pyzmq==26.2.1
rasterio==1.3.11
redis==6.1.1
referencing==0.35.1
requests==2.32.3
requests-unixsocket==0.2.0
rest-framework-simplejwt==0.0.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.20.1
rsa==4.9.1
s3fs==2024.10.0
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
SecretStorage==2.3.1
selenium==4.27.1
Send2Trash==1.8.3
simplejson==3.16.0
six==1.14.0
sniffio==1.3.1
snuggs==1.4.7
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.3
stack-data==0.6.3
systemd-python==234
terminado==0.18.1
texttable==1.6.2
threadpoolctl==3.5.0
tinycss2==1.2.1
tinynetrc==1.3.1
tomli==2.2.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
trio==0.27.0
trio-websocket==0.12.2
typing-extensions==4.12.2
tzdata==2025.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==2.2.3
usb-creator==0.3.7
uvicorn==0.33.0
vine==5.1.0
wadllib==1.3.3
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
werkzeug==3.0.6
wrapt==1.17.2
wsproto==1.2.0
xarray==2023.1.0
xgboost==2.1.4
xkit==0.0.0
yarl==1.15.2
zipp==3.20.2
